// src/components/buttons/buttons.interface.tsx
export type ButtonColorType =
  | "inherit"
  | "primary"
  | "secondary"
  | "success"
  | "error"
  | "warning"
  | "info";

export type ButtonVariantType = "text" | "outlined" | "contained";

export type ButtonSizeType = "small" | "medium" | "large";

export type ButtonTypeType = "submit" | "button" | "reset";
// ...
export interface CommonLoadingButtonPropsI {
  onClick?: (...arg: any) => void;
  children: ReactNode;
  color?: ButtonColorType;
  customStyles?: any | any;
  disabled?: boolean;
  loading?: boolean;
  size?: ButtonSizeType;
  type?: ButtonTypeType;
  variant?: ButtonVariantType;
  className?: string;
  primary?: boolean;
  fullWidth?: boolean;
  startIcon?: null | ReactNode;
  buttonLabel?: string;
}